Ejercicio14, corregido

+++ Ejercicio14,11/Program.cs
@@ -5,9 +5,9 @@ int a = int.Parse(Console.ReadLine());
 Console.WriteLine("Ingrese el divisor:_ ");
 int b = int.Parse(Console.ReadLine());
 
-int resultado = Divicion(a, b);
+double resultado = Divicion(a, b);
 Console.WriteLine("El resultado es:_ " + resultado);
-static int Divicion(int a, int b)
+static double Divicion(int a, int b)
 {
     return a / b;
 }
